@@ -8,27 +8,26 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "awobaz/compoships": "^1.1",
+        "php": "^7.3",
         "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^1.0",
-        "guzzlehttp/guzzle": "^6.3",
-        "laravel-notification-channels/webpush": "^5.0",
-        "laravel/framework": "^7.0",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "inertiajs/inertia-laravel": "^0.3.3",
+        "jantinnerezo/livewire-alert": "^2.0.2",
+        "jenssegers/date": "^4.0",
+        "laravel-frontend-presets/tall": "^2.1",
+        "laravel/framework": "^8.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
-        "laravelcollective/html": "^6.1",
-        "maatwebsite/excel": "^3.1",
-        "rachidlaasri/laravel-installer": "^4.1",
-        "spatie/laravel-cookie-consent": "^2.12",
-        "spatie/laravel-feed": "^2.7"
+        "livewire/livewire": "^2.3",
+        "mattlibera/livewire-flash": "^0.4.0"
     },
     "require-dev": {
-        "facade/ignition": "^2.0",
+        "barryvdh/laravel-debugbar": "^3.5",
+        "facade/ignition": "^2.3.6",
         "fzaninotto/faker": "^1.9.1",
         "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3"
     },
     "config": {
         "optimize-autoloader": true,
@@ -42,22 +41,16 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "files": [
-            "app/helpers.php"
-        ],
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
             "Tests\\": "tests/"
         }
     },
-   
     "minimum-stability": "dev",
     "prefer-stable": true,
     "scripts": {
