@@ -2,6 +2,17 @@
 
 namespace App\Providers;
 
+use App\Events\ClubCreated;
+use App\Events\ClubOwnerCreated;
+use App\Events\UserCreated;
+use App\Listeners\ClearClubIdFromSession;
+use App\Listeners\FirstTimeLogin;
+use App\Listeners\RegisterClubConfig;
+use App\Listeners\RegisterClubOwner;
+use App\Listeners\RegisterSettings;
+use App\Listeners\SetClubIdInSession;
+use Illuminate\Auth\Events\Login;
+use Illuminate\Auth\Events\Logout;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
@@ -18,6 +29,22 @@ class EventServiceProvider extends ServiceProvider
         Registered::class => [
             SendEmailVerificationNotification::class,
         ],
+        ClubCreated::class => [
+            RegisterClubOwner::class,
+        ],
+        UserCreated::class => [
+            RegisterSettings::class,
+        ],
+        ClubOwnerCreated::class => [
+            RegisterClubConfig::class,
+        ],
+        Login::class => [
+            SetClubIdInSession::class,
+            FirstTimeLogin::class,
+        ],
+        Logout::class => [
+            ClearClubIdFromSession::class,
+        ]
     ];
 
     /**
@@ -27,8 +54,6 @@ class EventServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        parent::boot();
-
         //
     }
 }
