@@ -1,80 +1,88 @@
 <!doctype html>
 <html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
-<head>
-   <meta name="viewport" content="width=device-width, initial-scale=1">
 
+<head>
+
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <!-- CSRF Token -->
     <meta name="csrf-token" content="{{ csrf_token() }}">
 
-    <title>{{ config('app.name', 'Laravel') }}</title>
+    <title>{{ config('app.name', 'KeizerPHP')}}</title>
 
     <!-- Scripts -->
     <script src="{{ asset('js/app.js') }}" defer></script>
-    <link href="//netdna.bootstrapcdn.com/bootstrap/3.0.0/css/bootstrap.min.css" rel="stylesheet">
-    <script src="https://code.jquery.com/jquery-2.0.3.min.js"></script> 
-    <script src="//netdna.bootstrapcdn.com/bootstrap/3.0.0/js/bootstrap.min.js"></script>
-
-    <!-- Fonts -->
-    <link rel="dns-prefetch" href="//fonts.gstatic.com">
-    <link href="https://fonts.googleapis.com/css?family=Nunito" rel="stylesheet">
-
+    <script src="https://kit.fontawesome.com/684ea5eff4.js" crossorigin="anonymous"></script>
+    <script src="https://unpkg.com/turbolinks@5.2.0"></script>
     <!-- Styles -->
+    @livewireStyles
     @auth
-    @if(settings()->has('layout'))
-        @if(settings()->get('layout') === 'app')
-            <link href="{{ asset('css/app.css') }}" rel="stylesheet">
-        @elseif(settings()->get('layout') === 'blue')
-        <link href="{{ asset('css/blue.css') }}" rel="stylesheet">
-       @endif
-    @else    
-        <link href="{{ asset('css/app.css') }}" rel="stylesheet">
+    @if(config('app.layout', 'app'))
+    @if(config('app.layout') === 'app')
+    <link href="{{ asset('css/app.css') }}" rel="stylesheet">
+    @elseif(config('app.layout') === 'blue')
+    <link href="{{ asset('css/blue.css') }}" rel="stylesheet">
+    @endif
+    @else
+    <link href="{{ asset('css/app.css') }}" rel="stylesheet">
     @endif
     @endauth
     @guest
-        <link href="{{ asset('css/app.css') }}" rel="stylesheet">
-        @endguest
-
-    <!-- Editable -->
-    
-    <script src="//cdnjs.cloudflare.com/ajax/libs/x-editable/1.5.0/bootstrap3-editable/js/bootstrap-editable.js"></script>
-    <link href="//cdnjs.cloudflare.com/ajax/libs/x-editable/1.5.0/bootstrap3-editable/css/bootstrap-editable.css" rel="stylesheet"/>
-    <script>
-     $(document).ready(function(){
-         setInterval(function(){
-      $("#funky").load(window.location.href + " #funky" );
-}, 10000);   
-    });
-    </script>
-
-    <!-- DataTable -->
-    <script src="https://cdn.datatables.net/1.10.21/js/jquery.dataTables.min.js"></script>
-    <script src="https://cdn.datatables.net/1.10.21/js/dataTables.bootstrap4.min.js"></script>
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.1.3/css/bootstrap.css" rel="stylesheet"/>
-    <link href="https://cdn.datatables.net/1.10.21/css/dataTables.bootstrap4.min.css" rel="stylesheet"/>
+    <link href="{{ asset('css/app.css') }}" rel="stylesheet">
+    @endguest
 </head>
-<body>
 
-    @include('inc.navbar')
-    <div class="container">
-        <div class="row justify-content-center">
-            <div class="col-md-8">
-                
-                @include('inc.messages')
-                @yield('content')
-            
+<body class="font-sans leading-normal tracking-normal mt-12">
+    @include('layouts.navbar')
+    <div class="flex flex-col md:flex-row">
+        @include('layouts.sidebar')
+        <div class="main-content flex-1 bg-gray-100 mt-12 md:mt-2 pb-24 md:pb-5">
+            <div class="flex flex-wrap">
+                <div class="w-full h-screen p-6 overflow-auto">
+                    @if (session()->has('message'))
+                        <div class="bg-yellow-lightest border-t-4 border-yellow rounded-b text-yellow-darkest px-4 py-3 shadow-md my-2" role="alert">
+                            <div class="flex">
+                                <div>
+                                    {{ session('message') }}
+                                </div>
+                            </div>
+                        </div>
+                    @elseif(session()->has('success'))
+                        <div class="bg-teal-lightest border-t-4 border-teal rounded-b text-teal-darkest px-4 py-3 shadow-md my-2" role="alert">
+                            <div class="flex">
+                                <div>
+                                    {{ session('success') }}
+                                </div>
+                            </div>
+                        </div>
+                    @elseif(session()->has('error'))
+                        <div class="bg-red-lightest border-t-4 border-red rounded-b text-red-darkest px-4 py-3 shadow-md my-2" role="alert">
+                            <div class="flex">
+                                <div>
+                                    {{ session('error') }}
+                                </div>
+                            </div>
+                        </div>
+                    @endif
+                    @yield('content')
+                </div>
             </div>
-        
         </div>
     </div>
-@auth
-@if(settings()->has('notifications'))
-    @if(settings()->get('notifications') == 2 OR settings()->get('notifications') == 3)
+
+    @auth
+    @if(config('app.notifications'))
+    @if(config('app.notifications') == 2 OR config('app.notifications') == 3)
     <script src="{{ asset('js/enable-push.js') }}" defer></script>
     @endif
-@endif
-@endauth  
+    @endif
+    @endauth
+    @livewireScripts
+    <script src="https://cdn.jsdelivr.net/gh/livewire/turbolinks@v0.1.x/dist/livewire-turbolinks.js"
+        data-turbolinks-eval="false"></script>
+    <x-livewire-alert::scripts />
+
 </body>
 
 </html>
-
-                      
