@@ -1,4 +1,6 @@
-const mix = require('laravel-mix');
+const mix = require("laravel-mix");
+
+require("laravel-mix-tailwind");
 
 /*
  |--------------------------------------------------------------------------
@@ -11,10 +13,11 @@ const mix = require('laravel-mix');
  |
  */
 
-mix.autoload({
-    jquery: ['$', 'window.jQuery', 'jQuery'],
-    'popper.js/dist/umd/popper.js': ['Popper']
-    })
-    .js('resources/js/app.js', 'public/js')
-    .sass('resources/sass/app.scss', 'public/css');
-    
+mix.js("resources/js/app.js", "public/js/app.js")
+    .sass("resources/sass/app.scss", "public/css/app.css")
+    .tailwind("./tailwind.config.js")
+    .sourceMaps();
+
+if (mix.inProduction()) {
+    mix.version();
+}
